Refactor iOS AI chat store tests
