+++ apps/ios/Flashcards/FlashcardsTests/AI/Store/Support/AIChatStoreTestSupport.swift
@@ -1,0 +1,2 @@
+enum AIChatStoreTestSupport {
+}
